put inverses in same file as presentatons
